changed "an" to "a" on line 304
line 304

was: <p>The weight for each step is determined by a "halting neuron". It's a sigmoid neuron that looks at the RNN state and gives an halting weight, which we can think of as the probability that we should stop at that step.</p>

proposed change: <p>The weight for each step is determined by a "halting neuron". It's a sigmoid neuron that looks at the RNN state and gives a halting weight, which we can think of as the probability that we should stop at that step.</p>